back/app/requirements.txt
fastapi
uvicorn[standard]
influxdb-client
influxdb3-python
playwright
pillow

pydantic>=2
pyarrow 

httpx
cachetools
pandas